pyproject.toml
[project]
name = "agi-1"
version = "0.1.0"
description = "Web browsing agent service using Brave Search, HTTP fetch, and Playwright"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "trafilatura>=1.6.0",
    "diskcache>=5.6.0",
    "playwright>=1.48.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agi"]

